add test cases for url columns
